+++ src/common/src/ConfirmDialog/ConfirmDialog.tsx
@@ -1,51 +1,56 @@
 import { Heading } from '@amsterdam/asc-ui';
-import React from 'react';
+import React, { useEffect } from 'react';
+import { BehaviorSubject } from 'rxjs';
 import Modal from '../Modal/Modal';
 import { ModalBlockStyle } from '../Modal/ModalStyles';
 import { ButtonStyles } from './ConfirmDialogStyles';
-import { BehaviorSubject } from 'rxjs';
 
 export interface IState {
-	title: string;
 	message: string;
-	textCancelButton: string;
-	textConfirmButton: string;
-	onCancel: () => void;
 	onConfirm: () => void;
+	title?: string;
+	textCancelButton?: string;
+	textConfirmButton?: string;
+	onCancel?: () => void;
 }
 
-const defaultMessage = 'Weet u zeker dat u dit item definitief wilt verwijderen?';
 const initialState: IState = {
-	title: 'Waarschuwing',
 	message: '',
-	textCancelButton: 'Nee',
-	textConfirmButton: 'Ja',
-	onCancel: () => {},
 	onConfirm: () => {},
 };
 
 const store = new BehaviorSubject(initialState);
 
-export const confirm = ({ ...props }: Partial<IState>) => {
-	store.next({ ...initialState, message: defaultMessage, ...props });
+export const confirm = ({
+	title = 'Waarschuwing',
+	message = 'Weet u zeker dat u dit item definitief wilt verwijderen?',
+	textConfirmButton = 'Ja',
+	textCancelButton = 'Nee',
+	onConfirm = () => {},
+	onCancel,
+}: IState) => {
+	store.next({ title, message, textConfirmButton, textCancelButton, onCancel, onConfirm });
 };
 
-const ConfirmDialog: React.FC<{}> = () => {
+const ConfirmDialog: React.FC = () => {
 	const [state, setState] = React.useState<IState>(initialState);
 	const [isVisible, setIsVisible] = React.useState<boolean>(false);
 
-	React.useEffect(() => {
-		store.subscribe(({ ...props }) => {
+	useEffect(() => {
+		store.subscribe((props) => {
 			setState({ ...props });
 			if (props.message) {
 				setIsVisible(true);
 			}
 		});
-		return () => setState(initialState);
+		return () => {
+			setIsVisible(false);
+			store.next(initialState);
+		};
 	}, []);
 
 	return (
-		<Modal id="confirm-dialog" open={isVisible}>
+		<Modal id="confirm-dialog" open={isVisible} size="sm">
 			<Modal.TopBar hideCloseButton>
 				<Heading forwardedAs="h4">{state.title}</Heading>
 			</Modal.TopBar>
@@ -55,27 +60,25 @@ const ConfirmDialog: React.FC<{}> = () => {
 			<Modal.Actions>
 				<Modal.Actions.Left>
 					<ButtonStyles
-						data-testid="cancel-button"
-						style={{ marginRight: '8px' }}
-						variant="primaryInverted"
-						onClick={() => {
-							setIsVisible(false);
-							if (typeof state.onCancel !== 'undefined') {
-								state.onCancel();
-							}
-						}}
-					>
-						{state.textCancelButton}
-					</ButtonStyles>
-					<ButtonStyles
 						data-testid="confirm-button"
 						variant="primary"
+						style={{ marginRight: '8px' }}
 						onClick={() => {
 							setIsVisible(false);
 							state.onConfirm();
 						}}
 					>
 						{state.textConfirmButton}
+					</ButtonStyles>
+					<ButtonStyles
+						data-testid="cancel-button"
+						variant="primaryInverted"
+						onClick={() => {
+							setIsVisible(false);
+							state.onCancel && state.onCancel();
+						}}
+					>
+						{state.textCancelButton}
 					</ButtonStyles>
 				</Modal.Actions.Left>
 			</Modal.Actions>