+++ src/common/src/DocumentTable/DocumentTable.stories.tsx
@@ -9,7 +9,7 @@ export default {
 };
 
 const Template: Story<React.ComponentProps<typeof DocumentTable>> = (args) => {
-	return <DocumentTable {...args} pageSize={3} />;
+	return <DocumentTable {...args} />;
 };
 
 export const Default = Template.bind({});
@@ -17,4 +17,22 @@ Default.args = {
 	rows: documents,
 	disableFilterRow: false,
 	disableRemoval: false,
+	pageSize: 3,
+};
+
+export const NoFilter = Template.bind({});
+NoFilter.args = {
+	rows: documents,
+	disableFilterRow: true,
+	disableRemoval: false,
+	pageSize: 4,
+};
+
+export const Loading = Template.bind({});
+Loading.args = {
+	rows: [],
+	disableFilterRow: false,
+	disableRemoval: false,
+	loading: true,
+	pageSize: 3,
 };